live4/pyproject.toml
[tool.poetry]
name = "lives"
version = "0.1.0"
description = ""
authors = ["Guilherme Pupio <guicpupio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^20.5"}


[tool.poetry.group.dev.dependencies]
gnureadline = "^8.1.2"


[tool.poetry.group.live2.dependencies]
pikepdf = "^8.4.0"
requests = "^2.31.0"
django = "^4.2.4"
psutil = "^5.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
